pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "transcritor"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
]

[project.optional-dependencies]
transcription = [
    "openai-whisper @ git+https://github.com/openai/whisper.git@517a43ecd132a2089d85f4ebc044728a71d49f6e",
    "torch==2.5.1",
    "moviepy==2.1.2",
    "sounddevice==0.5.1",
    "wavio==0.0.9",
]
api = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "celery[redis]>=5.3",
    "httpx>=0.27",
    "python-multipart>=0.0.9",
]
cli = [
    "click>=8.1",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "httpx>=0.27",
]

[project.scripts]
transcritor = "transcritor.cli.commands:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "slow: tests that load the real Whisper model",
    "e2e: end-to-end tests requiring Docker environment",
]

[tool.coverage.run]
source = ["src/transcritor"]
omit = ["*/cli/*"]
